+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased (v0.5.1)
 * Change encoded size prediction logic to estimate video stream size only. Change text to clarify this.
 * Improve video size prediction logic to account for samples that do not turn out as 20s.
+* Fix ffmpeg output size parsing kB inaccuracy.
 
 # v0.5.0
 * Default to .mkv output format for all inputs (except .mp4 which will continue to output .mp4 by default).