+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 * Improve video size prediction logic to account for samples that do not turn out as 20s.
 * Fix full-pass sample encode progress bar.
 * Use label "Full pass" instead of "Sample 1/1" when doing a full pass _sample-encode_.
+* Add VMAF auto model, n_threads & scaling documentation.
 
 # v0.5.0
 * Default to .mkv output format for all inputs (except .mp4 which will continue to output .mp4 by default).
-  This also applies to ffmpeg encoder sample output format. The previous behavior used the input extension
+  This also applies to ffmpeg encoder sample output format. The previous behaviour used the input extension
   which may not have supported av1 (e.g. .m2ts).
 * For _auto-encode_ use the output extension also for ffmpeg encoder sample outputs if applicable.
 * When creating lossless samples for encode analysis use .mkv (or .mp4) extension for better ffmpeg compatibility.